Clean IslandHook Ne plus utiliser org.bukkit.Player dans les fonctions SkyblockManager et IslandHook Début de l'ajout des futures commandes Changement du updateRole en updatePlayers Le clear du joueur au moment de supprimer l'ile à été temporairement enlevé le temps de terminer le reste du code

+++ src/main/java/fr/euphyllia/skyfolia/commands/subcommands/DeleteSubCommand.java
@@ -3,6 +3,7 @@ package fr.euphyllia.skyfolia.commands.subcommands;
 import fr.euphyllia.skyfolia.Main;
 import fr.euphyllia.skyfolia.api.event.SkyblockRemoveEvent;
 import fr.euphyllia.skyfolia.api.skyblock.Island;
+import fr.euphyllia.skyfolia.api.skyblock.Players;
 import fr.euphyllia.skyfolia.commands.SubCommandInterface;
 import fr.euphyllia.skyfolia.configuration.ConfigToml;
 import fr.euphyllia.skyfolia.configuration.section.WorldConfig;
@@ -12,7 +13,6 @@ import fr.euphyllia.skyfolia.utils.WorldEditUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.bukkit.Bukkit;
-import org.bukkit.GameMode;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
@@ -20,6 +20,8 @@ import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.List;
+import java.util.concurrent.Executors;
+import java.util.concurrent.ScheduledExecutorService;
 
 public class DeleteSubCommand implements SubCommandInterface {
 
@@ -31,37 +33,34 @@ public class DeleteSubCommand implements SubCommandInterface {
             if (!(sender instanceof Player player)) {
                 return true;
             }
-            player.setGameMode(GameMode.SPECTATOR);
-            Bukkit.getAsyncScheduler().runNow(plugin, scheduledTask -> {
-                SkyblockManager skyblockManager = plugin.getInterneAPI().getSkyblockManager();
-                Island island = skyblockManager.getIslandByOwner(player).join();
-                if (island == null) {
-                    // pas d'ile
-                    return;
-                }
+            ScheduledExecutorService executor = Executors.newSingleThreadScheduledExecutor();
+            try {
+                executor.execute(() -> {
+                    SkyblockManager skyblockManager = plugin.getInterneAPI().getSkyblockManager();
+                    Island island = skyblockManager.getIslandByOwner(player).join();
+                    if (island == null) {
+                        // pas d'ile
+                        return;
+                    }
+                    if (!island.getOwnerId().equals(player.getUniqueId())) {
+                        return; // Seul l'owner peut !
+                    }
 
-                island.setDisable(true);
-                PlayerUtils.teleportPlayerSpawn(plugin, player);
-                for (WorldConfig worldConfig : ConfigToml.worldConfigs) {
-                    WorldEditUtils.deleteIsland(plugin, island, Bukkit.getWorld(worldConfig.name()), 50);
-                }
-                player.getScheduler().run(plugin, scheduledTask1 -> {
-                    if (ConfigToml.clearInventoryWhenDeleteIsland) {
-                        player.getInventory().clear();
-                    }
-                    if (ConfigToml.clearEnderChestWhenDeleteIsland) {
-                        player.getEnderChest().clear();
-                    }
-                    if (ConfigToml.clearEnderChestWhenDeleteIsland) {
-                        player.setTotalExperience(0);
-                        player.sendExperienceChange(0, 0); // Mise à jour du packet
-                    }
-                    player.setGameMode(GameMode.SURVIVAL);
-                }, null);
+                    this.teleportPlayerSpawn(plugin, island);
 
-                SkyblockRemoveEvent skyblockRemoveEvent = new SkyblockRemoveEvent(island, player.getUniqueId());
-                Bukkit.getServer().getPluginManager().callEvent(skyblockRemoveEvent);
-            });
+                    island.setDisable(true);
+
+                    for (WorldConfig worldConfig : ConfigToml.worldConfigs) {
+                        WorldEditUtils.deleteIsland(plugin, island, Bukkit.getWorld(worldConfig.name()), 50);
+                    }
+
+                    SkyblockRemoveEvent skyblockRemoveEvent = new SkyblockRemoveEvent(island);
+                    Bukkit.getServer().getPluginManager().callEvent(skyblockRemoveEvent);
+                });
+            } finally {
+                executor.shutdown();
+            }
+
             return true;
         } catch (Exception ex) {
             logger.fatal("Suppression ile a un bug", ex);
@@ -73,4 +72,30 @@ public class DeleteSubCommand implements SubCommandInterface {
     public @Nullable List<String> onTabComplete(@NotNull Main plugin, @NotNull CommandSender sender, @NotNull Command command, @NotNull String label, @NotNull String[] args) {
         return null;
     }
+
+    private void teleportPlayerSpawn(Main plugin, @NotNull Island island) {
+        for (Players players : island.getMembers()) {
+            Player member = Bukkit.getPlayer(players.getMojangId());
+            if (member == null) continue;
+            if (member.isOnline()) {
+                PlayerUtils.teleportPlayerSpawn(plugin, member);
+            }
+        }
+    }
+
+    /* Todo Va être déplacer
+    player.setGameMode(GameMode.SPECTATOR);
+        if (ConfigToml.clearInventoryWhenDeleteIsland) {
+            player.getInventory().clear();
+        }
+        if (ConfigToml.clearEnderChestWhenDeleteIsland) {
+            player.getEnderChest().clear();
+        }
+        if (ConfigToml.clearEnderChestWhenDeleteIsland) {
+            player.setTotalExperience(0);
+            player.sendExperienceChange(0, 0); // Mise à jour du packet
+        }
+
+        player.setGameMode(GameMode.SURVIVAL);
+     */
 }
